Removed Cobblemon UI Tweaks Features
Updated LICENSE as it no longer uses Landon Wills code
