+++ LICENSE
@@ -1,6 +1,5 @@
 MIT License
 
-Copyright (c) 2024 Landon Wills
 Copyright (c) 2024 RIze2kNight
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
