+++ README.md
@@ -1,17 +1,14 @@
 # cobblemonuitweaks-1.21.1
-A port of Cobblemon UI Tweaks by Landon-Wills/Landojw for 1.6 Cobblemon snapshot
-Modified under the terms of the MIT License.
+A QoL mod for cobblemon 1.6 snapshot
 
-Credit goes to Landojw for the original code
-Original Repository: https://github.com/landon-wills/cobblemon-ui-tweaks
+Features(Yay I'm not a bum):
+- Edit PC Box number to jump to it (Press Enter to jump)
+- Hidden Ability Highlighter in PC and Summary
 
-OG Features refactored to 1.21.1:
+Removed OG Features of cobblemonuitweaks to 1.21.1:
 - Using the summary keybind when hovering a Pokemon in a PC or pasture will open it's summary screen
-- PCs and pastures will open the last viewed box when re-opened 
+- PCs and pastures will open the last viewed box when re-opened
 - Scrolling in a PC or pasture will move to the next/previous box
 - Pokemon's portrait in battle, a tooltip will now render showing the Pokemon's species, form, types, and trainer.
 - Fixes type icons appearing larger than they should on enemy portrait tooltip
 
-Added Features(Yay I'm not a bum):
-- Edit PC Box number to jump to it (Press Enter to jump)
-- Hidden Ability Highlighter in PC and Summary
