+++ gradle.properties
@@ -10,7 +10,7 @@ loader_version=0.16.9
 fabric_kotlin_version=1.12.3+kotlin.2.0.21
 
 # Mod Properties
-mod_version=1.0.0
+mod_version=1.1.0
 maven_group=com.rize2knight
 archives_base_name=cobblemonuitweaks
